Fix and update

+++ models/package.go
@@ -1,7 +1,7 @@
 package models
 
 type PackageData struct {
-	Weight  float32 `json:"weight"`
+	Weight  float64 `json:"weight"`
 	Size    string  `json:"size"`
 	Service string  `json:"service,omitempty" default:"Standard"`
 }
